Remove excess `stricmp` definitions

+++ czip/BitIO.cpp
@@ -34,16 +34,17 @@
  * $NoKeywords: $
  */
 
+#include <algorithm>
+#include <cstdio>
 #include <cstdint>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
+#include <cstdlib>
+#include <cstring>
+#include <cerrno>
+
 #include <sys/stat.h>
 #include <errno.h>
 #include <Macros.h>
 #include "CZip.h"
-
-#include <algorithm>
 
 BITFILE *CZip::OpenInputBitFile(char *filename) {
   BITFILE *bit_file;
@@ -172,9 +173,9 @@ int CZip::InputBit(BITFILE *bfile) {
   return (value ? 1 : 0);
 }
 
-uint32_t CZip::InputBits(BITFILE *bfile, int bitcount) {
-  uint32_t mask;
-  uint32_t return_value;
+ulong CZip::InputBits(BITFILE *bfile, int bitcount) {
+  ulong mask;
+  ulong return_value;
 
   mask = 1L << (bitcount - 1);
   return_value = 0;
